tools-ma: fix problem with command size in check_file_utils.py
The method that gets all statements with specific names uses grep
to get the lines where names appears. The issue is that the grep
command was getting too big and exceeding the command limit from OS
(specially in x86). This patch split the names searched and run grep
with a small set to avoid the problem.

Change-Id: I01fccc4a3287ece87ffd6d2bb83d9a3e1100cd52
Signed-off-by: Roberto Oliveira <rdutra@linux.vnet.ibm.com>