tools-ma: make method that get ifdef blocks generic and move to core
Make method that get ifdefs, from performance degradation checker,
more generic, move it to MA core and addapt performance degradation
checker to use this method.
This changes ended up fixing a bug in performance degradation that was
reporting as problem when a PPC part appeared before a x86 part in a
ifdef block.

Change-Id: I3ab6a62f02e84ce2b15794b4ad7fa72299258962
Signed-off-by: Roberto Oliveira <rdutra@linux.vnet.ibm.com>