Add `--verbose` option
Support a `--verbose` / `-v` option which will emit more information.

Use the verbose option to emit information about errors and warnings
generated during clang indexing.

Pass `argv` instead of individual members to `_run_checker`.

(Change a uses of `args` to `argv` to better represent its purpose and
avoid a conflict with a local variable.)

The new option is global, rather than associated with a subcommand, so
the command syntax would be, for example:
```
$ ma --verbose run -c asm .
```
Signed-off-by:  Paul A. Clarke <pc@us.ibm.com>